Move evidenceTypeHelp from Increase to All Claims
